Restructure complete. Tidied up a few files, all in working order
