+++ .babelrc
@@ -1,0 +1,6 @@
+{
+  "presets": ["babel-preset-env", "react"],
+  "plugins": [
+    ["transform-class-properties", { "spec": true }]
+  ]
+}
