+++ package.json
@@ -5,12 +5,12 @@
   "private": true,
   "dependencies": {
     "axios": "^0.17.1",
-    "babel-cli": "^6.26.0",
-    "eslint-plugin-react": "^7.5.1",
+    "babel-preset-env": "^1.6.1",
     "gh-pages": "^1.1.0",
     "leaflet": "^1.2.0",
     "prop-types": "^15.6.0",
     "react": "^16.2.0",
+    "react-addons-test-utils": "^15.6.2",
     "react-bootstrap": "^0.32.0",
     "react-dom": "^16.2.0",
     "react-fontawesome": "^1.6.1",
@@ -18,7 +18,8 @@
     "react-router": "^4.2.0",
     "react-router-dom": "^4.2.2",
     "react-scripts": "1.1.0",
-    "webpack": "^3.10.0"
+    "react-test-renderer": "^16.2.0",
+    "sinon": "^4.2.0"
   },
   "scripts": {
     "predeploy": "npm run build",
@@ -31,10 +32,21 @@
   "devDependencies": {
     "babel-core": "^6.26.0",
     "babel-jest": "^22.0.1",
+    "enzyme": "^3.3.0",
+    "enzyme-adapter-react-16": "^1.1.1",
     "eslint": "^4.13.0",
     "eslint-config-google": "^0.9.1",
     "eslint-plugin-react": "^7.5.1",
+    "expect": "^22.1.0",
+    "identity-obj-proxy": "^3.0.0",
     "jest": "^22.0.1",
     "regenerator-runtime": "^0.11.1"
+  },
+  "jest": {
+    "moduleNameMapper": {
+      "\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/__mocks__/fileMock.js",
+      "\\.(css|less)$": "identity-obj-proxy"
+    },
+    "setupTestFrameworkScriptFile": "<rootDir>/test-setup.js"
   }
 }
