+++ src/components/MapResults.js
@@ -1,33 +1,8 @@
-import React from 'react';
-import {  Map, TileLayer, Marker, Popup } from 'react-leaflet';
+import React, { Component } from 'react';
+import ServiceMapMarker from './ServiceMapMarker';
+import {  Map, TileLayer } from 'react-leaflet';
 
-class ServiceMapMarker extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {};
-  }
-  render() {
-    if (this.props.record.LATITUDE && this.props.record.LONGITUDE) {
-      let position = [this.props.record.LATITUDE, this.props.record.LONGITUDE];
-      return (
-        <Marker position={position}>
-          <Popup>
-            <span>
-              <h5>{this.props.record.PROVIDER_NAME}</h5>
-              <h6>{this.props.record.SERVICE_NAME}</h6>
-              <p>{this.props.record.SERVICE_DETAIL}</p>
-              <p>{this.props.record.PHYSICAL_ADDRESS}</p>
-            </span>
-          </Popup>
-        </Marker>
-      );
-    }
-    else {
-      return '';
-    }
-  }
-}
-class MapResults extends React.Component {
+class MapResults extends Component {
   constructor(props) {
     super(props);
     if(this.props.longitude && this.props.latitude) {
@@ -61,5 +36,4 @@ class MapResults extends React.Component {
     );
   }
 }
-
 export default MapResults;
