+++ src/components/SearchForm.js
@@ -1,28 +1,24 @@
-import React from 'react';
+import React, { Component } from 'react';
 import { Button, FormGroup, FormControl, InputGroup } from 'react-bootstrap';
 import FontAwesome from 'react-fontawesome';
-import AddressResolver from './addressresolver.jsx';
+import AddressResolver from './AddressResolver';
 
-class SearchForm extends React.Component {
+class SearchForm extends Component {
   constructor(...args) {
     super(...args);
     this.state = {keyword: '', address: '', selected_address: '', longitude: '', latitude: '', results: []};
-    this.handleKeywordChange = this.handleKeywordChange.bind(this);
-    this.handleAddress = this.handleAddress.bind(this);
-    this.handleAddressSelection = this.handleAddressSelection.bind(this);
-    this.handleSubmit = this.handleSubmit.bind(this);
   }
-  handleKeywordChange(event) {
+  handleKeywordChange = event => {
     this.setState({keyword: event.target.value});
   }
-  handleAddress(event) {
+  handleAddress = event => {
     this.setState({address: event.target.value});
   }
-  handleAddressSelection(selected_address) {
+  handleAddressSelection = selected_address => {
     this.setState({selected_address: selected_address, address: selected_address.a, longitude: selected_address.x, latitude: selected_address.y});
     this.props.handler({selected_address: selected_address, longitude: selected_address.x, latitude: selected_address.y});
   }
-  handleSubmit(event) {
+  handleSubmit = event => {
     this.props.handler({keyword: this.state.keyword, location: this.state.location});
     event.preventDefault();
   }
