--- src/searchform.css
@@ -1,2 +1,0 @@
-.big {
-}
