+++ src/styles/index.css
@@ -4,6 +4,6 @@ body {
 }
 
 .leaflet-container {
-    height: 400px;
-    width: 100%;
+  height: 400px;
+  width: 100%;
 }
